requirements.txt
GPUtil==1.4.0
psutil==5.8.0
psycopg2-binary==2.9.2
